Combine backend and web logic into single exe
The backend

    runs.toml + chart-requests.toml -> charts.json

and frontend

    charts.json -> chart files

logic has been combined into a single haskell exe. That is, the backend
generates the json file, then shells out to npm to build the actual
wanted files. This greatly simplifies project organization and
distribution, as it means everything is combined in a single exe.

To do this, we use TH to embed the needed web files (e.g. js/ts),
and use the XDG cache dir as the build directory. We then copy the
final build outputs to the current working directory.