Fix Moment Ord
Ord Moment had a bug but that's not all. When implementing tests, we
discovered that Eq/Ord are unlawful. That is unfortunate because the
unlawful functionality (compare Timestamps by their 'common root') is
necessary for some operations (checking timestamp overlap, comparing
'moment's).

We should make Eq/Ord standard in the future, implement specific functions
for the above, and write property tests for them too.