Log exceptions before rethrowing
This is potentially redundant, though there is some benefit:

- If we ever log to a file, we'd want the fatal error there.

- Logging will include the callstack on "known exceptions", which is
possibly useful for debugging.

There is an _apparent_ benefit for GHC 9.10, but that merits
explanation. For some exceptions (e.g. file not found), it appears
that the exception is being truncated: i.e. logging produces a full
callstack, but the final error message is severly truncated i.e. stops
after throwIO in base.

The seems like an argument in favor of logging, since maybe the normal
exception handling is losing the callstack somehow? Notably, this only
happens on GHC 9.10 and for "known" exceptions.

That is, "known exceptions" are those we deliberately try to skip
callstacks since the callstacks are mostly unfriendly noise. In
GHC 9.12+, our ignore logic works fine, so actually no callstack is
printed at all (as we want). On GHC 9.10, however, the base callstack
is still printed because of internal changes to base's exception
machinery.

This gives the impression that we somehow lost a callstack that we
may care about. But this is misleading; we did not want such a
callstack anyway. It is only due to GHC 9.10's legacy exception
implementation that we saw a "partial" callstack at all.