Remove duration units
It turns out, we never used the fact that duration was polymorphic
over units i.e. could be seconds, minutes, and hours. We only ever
used seconds. Moreover, there are no plans to use these other units;
since parsing is the same regardless of unit, the fact that there can
be multiple units is entirely an internal detail.

We thus remove this capability as it significantly simplifies the
code.