Allow run labels with garmin
We allow the user to specify run labels in a separate file e.g.
run-labels.toml. The intention is to allow usage with garmin
activities.csv files, as the latter does not include any arbitrary
text fields. Because datetimes are unique, we can use them as a key in
the separate file, then combine all labels based on this key.