Replace toml with json(c)
The addition of chart-requests types has revealed an ugly truth:
Nested toml object/array syntax is pretty bad.

In particular, leaving off the 'period' on the 'sum' chart-type did
_not_ cause an error, presumably because the parse error causes the
key to be considered "unknown" (the field is optional), and unused
keys are not an error.

While such mistakes can occur with any typo, they are especially likely
when using nested array objects. Hence looking for a second option.

While json has the truly unfortunate property that comments are
invalid, microsoft's jsonc format allows them. We do not have an actual
parser for jsonc, but it is easy enough to write a parser that strips
comments ourselves, then hand the json off to aeson.