Only search for json config with 'chart' command
The config only affects the 'chart' command, hence it is a waste to
search for it otherwise. It also saves a useless 'Using config...'
log message for non-chart commands.

We leave it as part of the global options for now, as refactoring it
would take some work, and we may want to generalize the config in the
future e.g. setting preferred units.