Remove redundant chart title
Chart titles are redundant now that the selector exists, as it also
displays the title. We therefore remove chart title (which also helps with
vertical spacing), and add logic for truncating long titles, and displaying
the full title on hover.

This works pretty well. Alternatively, we could display the dropdown to the side
of the chart. We'd lose some horizontal space, but we'd gain vertical, and
the full title would be easier to read.