Implement chart command server
The chart command launches a server, rather than producing static
html files. The primary advantage is that we can statically generate
the frontend and dynamically load the user's charts.json file that
is generated by the backend.

In particular, this means the haskell backend no longer has to
run npm, so node is no longer a runtime dependency.

This also introduces the possibility of having an "end to end" test
e.g. launching the server and testing it with selenium. For now, we
have a simple sanity test that checks that the server correctly
launches.