Fix windows log encoding bug
The emoji was causing windows to crash, because apparently
windows + unicode + putStrLn is not handled correctly. It seems the
only robust course of action is to simply remove the emoji. This is
why we can't have nice things.

https://discourse.haskell.org/t/solved-what-the-range-of-characters-that-putstrln-support/8305/6

Separately, we replace the logger's usage of putText with putBinary,
because investigation revealed that using putText is silly. The
conversions are:

  1. Encode emoji literal as Text (fromString).
  2. Logger encodes to internal LogStr/Builder type (Text specifically
     encodes as UTF-8).
  3. Encode to ByteString.
  4. Decode _back_ to Text (as lenient UTF-8).
  5. Encode to String and send to putStrLn, where it presumably
     end back up as bytes (and involves system locale, which causes
     the crash).

Replacing putText with putBinary means we do not have to do steps
4 and 5. It also appears more robust, since switching to putBinary
"fixed" the error, in that we no longer got a crash; the output was
merely mangled, which is obviously better. Thus we make two changes
here:

1. Replace the emoji with an emoticon :-(
2. Use putBinary. This isn't the actual fix, though it's a sensible
   change regardless, and may actually be more robust anyway.