Make alias search return multiple hits
Previously, the alias search -- when specialized to List/Seq --
only took the first result even with multiple hits e.g.

  SearchFiles [Aliases [one, two]] == [one]

This means it does that Aliases does not distribute because we have

  SearchFiles [Aliases [one], Aliases [two]] == [one, two]

This contrasts with the other searchs:

  -- Aliases specialized to Maybe distributes
  SearchFiles [Aliases [one], Aliases [two]] == one
  SearchFiles [Aliases [one, two]] == one

  -- Infix and Aliases have the same Maybe semantics.
  SearchFiles [Infix one _, Infix two _] = one

  -- Infix and Aliases have the same List semantics.
  SearchFiles [Infix one _, Infix two _] = [one, two]

Finally, we already have a way to take the "first" result: Maybe.
Hence there is little reason to preserve this quirk. Thus we change
it for consistency and simplicity.

Note that this is purely an internal change, as there was nothing
using this combination:

  - Alias search with multiple aliases.
  - We are returning a List, not Maybe.
  - But we only want the first element from the List.

In fact, the only user of the List specialization is activities,
but we wanted all results, so we were already using

  SearchFiles [Aliases [one], Aliases [two]]

Which does what we want.