Simplify FileSearch type
We are no longer using the ability to search for 'multiple' files
simultaneously, so it is removed, for simplicity.

Also fix bug where aliases search would not find multiple files
when we had an exact match...but the semantics are such that
exact match and case-insense match should behave the same.