Fix docker static build
The static build failed at link time due to ld being unable to find
gmp. This is odd, since the installation of gmp and gmp-dev seems
fine. The error mentions something about being unable to find the
_static_ version, so maybe somehow the version of gmp that is
downloaded is dynamically linked?

In any case, fixing the alpine version to 3.21 appears to work
for now (CI was picking the new 3.22, with the error).