Improve chart param unit tests
We significantly improve the chart param unit tests. The primary
change is reducing the coupling between the tests. That is, prior
to this change, the tests' "doesFile/DirExist" logic was
centralized in the mocked effect, and then tests would choose
different hardcoded paths to get the behavior they want. This is
confusing, because to understand a particular test, we need to see
what paths it uses, then check the effect definition.

Now, we move all path behavior to the tests themselves. While this
is a bit more verbose, it makes the tests much more self-contained
and easier to understand.

We also make small syntactic improvements e.g. replacing verbose
optics usage with record setters or helper functions where
applicable.

Finally, we fix what is arguably a bug in the chart config logic.
That is, if the chart config is given an explicit path -- and it does
not exist -- it should cause an immediate error, not potentially try
the config data directory next. This is consistent with how other
paths work i.e. if an explicit path is given and does not exist,
it should throw an exception, according to its order in the priority
list.