Parse global type filters early
Global type filters e.g. filters like "type = X" that are defined
in the top-level 'filters' key in chart-requests, are now applied
during parsing.

That is, if we start to parse an activity (garmin or custom json), and
its type would be filtered out, we stop and move onto the next one.
In particular, this means that we will not unnecessarily report any
parse errors, if the activity would be filtered out anyway.

Note that this applies only to _global_ filters. Filters on individual
chart requests are only considered after parsing.

Implementation note: Because parsing now requires extra data (filters),
we remove instances in favor of manual parsing functions. This requires
some massaging -- especially to preserve error messages --
but is otherwise reasonably straightforward.