Hot-reload chart files
Allows hot-reloading of data files i.e. previously when a change was
made to some data file (e.g. activities.csv, chart-requests.jsonc),
the server would have to be re-run (pacer chart), and the page
refreshed.

Now all it requires is a page refresh, as the charts are created
on page load.